# client/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Manhwa Downloader API",
    "description": "Search manga sources, list chapters, and build downloadable archives.",
    "license": {
      "name": ""
    },
    "version": "0.2.0"
  },
  "paths": {
    "/chapters": {
      "get": {
        "tags": [
          "catalogue"
        ],
        "operationId": "chapters",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "description": "Source id or slug.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "query",
            "description": "Comic id as returned by search.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "description": "Chapter language for multi-language sources. Defaults to `en`.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChaptersResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "502": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/download": {
      "post": {
        "tags": [
          "downloads"
        ],
        "operationId": "start_download",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadAccepted"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Unsupported format on this server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "Too many downloads from this address; see Retry-After",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "501": {
            "description": "Source cannot download yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/download/cancel/{task_id}": {
      "post": {
        "tags": [
          "downloads"
        ],
        "operationId": "cancel",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cancelled"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Already finished",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/download/events/{task_id}": {
      "get": {
        "tags": [
          "downloads"
        ],
        "operationId": "events",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server-Sent Events stream of TaskStatus objects; closes on a terminal state",
            "content": {
              "text/event-stream": {}
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/download/file/{task_id}": {
      "get": {
        "tags": [
          "downloads"
        ],
        "operationId": "file",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The archive",
            "content": {
              "application/octet-stream": {}
            }
          },
          "404": {
            "description": "Unknown task, failed task, or an archive whose retention window has closed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Task still running",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/download/status/{task_id}": {
      "get": {
        "tags": [
          "downloads"
        ],
        "operationId": "status",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStatus"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Cancelled": {
        "type": "object",
        "required": [
          "status",
          "task_id"
        ],
        "properties": {
          "status": {
            "type": "string"
          },
          "task_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "ChapterRef": {
        "type": "object",
        "required": [
          "id",
          "number"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "number": {
            "type": "string",
            "description": "Chapter label used for file names (\"12\", \"12.5\")."
          }
        }
      },
      "ChaptersResponse": {
        "type": "object",
        "required": [
          "source",
          "comic_id",
          "volumes",
          "total_chapters"
        ],
        "properties": {
          "comic_id": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "total_chapters": {
            "type": "integer",
            "minimum": 0
          },
          "volumes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Volume"
            }
          }
        }
      },
...
      "DownloadAccepted": {
        "type": "object",
        "required": [
          "task_id",
          "state",
          "status",
          "message",
          "status_url",
          "events_url",
          "file_url"
        ],
        "properties": {
          "events_url": {
            "type": "string"
          },
          "file_url": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/TaskState"
          },
          "status": {
            "type": "string",
            "description": "Legacy field the old client displayed."
          },
          "status_url": {
            "type": "string"
          },
          "task_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "DownloadRequest": {
        "type": "object",
        "required": [
          "source",
          "chapters"
        ],
        "properties": {
          "chapters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChapterRef"
            }
          },
          "comic_title": {
            "type": "string"
          },
          "format": {
            "$ref": "#/components/schemas/Format"
          },
          "lang": {
            "type": [
              "string",
              "null"
            ]
          },
          "source": {
            "type": "string",
            "description": "Source id or slug."
          }
        }
      },
      "ErrorBody": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable code."
          },
          "message": {
            "type": "string"
          },
          "source": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "Format": {
        "type": "string",
        "enum": [
          "pdf",
          "cbz",
          "cbr",
          "epub"
        ]
      },
...
      "Volume": {
        "type": "object",
        "required": [
          "name",
          "chapters"
        ],
        "properties": {
          "chapters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Chapter"
            }
          },
          "name": {
            "type": "string",
            "description": "Display name, e.g. `\"Vol 1\"`. Sources without volumes use `\"Vol 1\"`."
          }
        }
      }
    }
  },
...
}